requirements.txt
appdirs==1.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
docker==5.0.0
filelock==3.0.12
Flask==1.1.2
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
nacos-sdk-python==0.1.6
pipenv==2020.11.15
pycparser==2.20
pywin32==227
requests==2.25.1
six==1.15.0
urllib3==1.26.4
virtualenv==20.4.6
virtualenv-clone==0.5.4
websocket-client==0.58.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
